Entire Rewrite
Turning what was the original PoC version of the site into the v2.0 beta for the site and rewriting the  site with a simple design to get the information publcily available first and will come back to the PoC design in the future. This is the very beginning of that redesign and the filing away of the PoC site.