Updating to Alpha 0.2.1
Made a fair number of changes to the site as well as added tons of new subpages and information to these pages to start bringing the site to a more finalised state. Added some info to the hardware subpage and added a ton of new pages for the  console firmwares and their respective pages.