Updating to Alpha 0.2.2
Updating the site to Alpha Build 0.2.2 with all of the currently needed subpages. These subpages are typically just placeholder pages for when I eventually begin to add the information to these pages. Also made some minor adjustments to the homepage as well as linking all of the buttons currently within the site.